refactor: rename rust ui bridge to host
