+++ yoyopod/ui/rust_host/__init__.py
@@ -1,0 +1,14 @@
+"""Rust UI host integration helpers."""
+
+from yoyopod.ui.rust_host.facade import RustUiFacade
+from yoyopod.ui.rust_host.protocol import UiEnvelope, UiProtocolError
+from yoyopod.ui.rust_host.snapshot import RustUiRuntimeSnapshot
+from yoyopod.ui.rust_host.supervisor import RustUiHostSupervisor
+
+__all__ = [
+    "RustUiFacade",
+    "RustUiHostSupervisor",
+    "RustUiRuntimeSnapshot",
+    "UiEnvelope",
+    "UiProtocolError",
+]
