+++ yoyopod/ui/rust_host/intents.py
@@ -1,0 +1,9 @@
+"""Intent payload helpers for the Rust UI host."""
+
+from __future__ import annotations
+
+from typing import Any
+
+RustUiIntentPayload = dict[str, Any]
+
+__all__ = ["RustUiIntentPayload"]
