+++ yoyopod/ui/rust_sidecar/__init__.py
@@ -1,6 +1,14 @@
-"""Rust UI PoC sidecar integration helpers."""
+"""Compatibility imports for the renamed Rust UI host bridge."""
 
-from yoyopod.ui.rust_sidecar.protocol import UiEnvelope, UiProtocolError
-from yoyopod.ui.rust_sidecar.state import RustUiRuntimeSnapshot
+from yoyopod.ui.rust_host import RustUiFacade, RustUiRuntimeSnapshot, UiEnvelope, UiProtocolError
+from yoyopod.ui.rust_host.facade import RustUiFacade as RustUiSidecarCoordinator
+from yoyopod.ui.rust_host.supervisor import RustUiHostSupervisor as RustUiSidecarSupervisor
 
-__all__ = ["RustUiRuntimeSnapshot", "UiEnvelope", "UiProtocolError"]
+__all__ = [
+    "RustUiFacade",
+    "RustUiRuntimeSnapshot",
+    "RustUiSidecarCoordinator",
+    "RustUiSidecarSupervisor",
+    "UiEnvelope",
+    "UiProtocolError",
+]
